langgraph_gemini_bot/requirements.txt
langchain
langgraph
google-generativeai
python-dotenv
mysql-connector-python  # Change this based on your database
